fixes issue #96
