+++ src/Automatic/Automatic.php
@@ -9,11 +9,9 @@ use Composer\Config;
 use Composer\Console\Application;
 use Composer\DependencyResolver\Operation\InstallOperation;
 use Composer\DependencyResolver\Operation\UpdateOperation;
-use Composer\DependencyResolver\Pool;
 use Composer\EventDispatcher\EventSubscriberInterface;
 use Composer\Factory;
 use Composer\Installer;
-use Composer\Installer\InstallerEvent;
 use Composer\Installer\InstallerEvents;
 use Composer\Installer\PackageEvent;
 use Composer\Installer\PackageEvents;
@@ -25,10 +23,7 @@ use Composer\Package\BasePackage;
 use Composer\Package\Locker;
 use Composer\Plugin\PluginEvents;
 use Composer\Plugin\PluginInterface;
-use Composer\Plugin\PreFileDownloadEvent;
-use Composer\Repository\ComposerRepository as BaseComposerRepository;
 use Composer\Repository\RepositoryFactory;
-use Composer\Repository\RepositoryInterface;
 use Composer\Repository\RepositoryManager;
 use Composer\Script\Event;
 use Composer\Script\ScriptEvents as ComposerScriptEvents;
@@ -46,7 +41,7 @@ use Narrowspark\Automatic\Installer\SkeletonInstaller;
 use Narrowspark\Automatic\Operation\Install;
 use Narrowspark\Automatic\Operation\Uninstall;
 use Narrowspark\Automatic\Prefetcher\ParallelDownloader;
-use Narrowspark\Automatic\Prefetcher\Prefetcher;
+use Narrowspark\Automatic\Prefetcher\PrefetcherTrait;
 use Narrowspark\Automatic\Prefetcher\TruncatedComposerRepository;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
@@ -58,6 +53,7 @@ class Automatic implements PluginInterface, EventSubscriberInterface
 {
     use ExpandTargetDirTrait;
     use GetGenericPropertyReaderTrait;
+    use PrefetcherTrait;
 
     /**
      * @var string
@@ -712,115 +708,6 @@ class Automatic implements PluginInterface, EventSubscriberInterface
 
         foreach ($jsonContents['scripts'][ScriptEvents::AUTO_SCRIPTS] as $cmd => $type) {
             $scriptExecutor->execute($type, $cmd);
-        }
-    }
-
-    /**
-     * Populate the provider cache.
-     *
-     * @param \Composer\Installer\InstallerEvent $event
-     *
-     * @return void
-     */
-    public function onPreDependenciesSolving(InstallerEvent $event): void
-    {
-        $listed   = [];
-        $packages = [];
-        $pool     = $event->getPool();
-        $pool     = \Closure::bind(function () {
-            foreach ($this->providerRepos as $k => $repo) {
-                $this->providerRepos[$k] = new class($repo) extends BaseComposerRepository {
-                    /**
-                     * A repository implementation.
-                     *
-                     * @var \Composer\Repository\RepositoryInterface
-                     */
-                    private $repo;
-
-                    /**
-                     * @param \Composer\Repository\RepositoryInterface $repo
-                     */
-                    public function __construct(RepositoryInterface $repo)
-                    {
-                        $this->repo = $repo;
-                    }
-
-                    /**
-                     * {@inheritdoc}
-                     */
-                    public function whatProvides(Pool $pool, $name, $bypassFilters = false)
-                    {
-                        $packages = [];
-
-                        if (! \method_exists($this->repo, 'whatProvides')) {
-                            return $packages;
-                        }
-
-                        foreach ($this->repo->whatProvides($pool, $name, $bypassFilters) as $k => $p) {
-                            $packages[$k] = clone $p;
-                        }
-
-                        return $packages;
-                    }
-                };
-            }
-
-            return $this;
-        }, clone $pool, $pool)();
-
-        foreach ($event->getRequest()->getJobs() as $job) {
-            if ($job['cmd'] !== 'install' || \strpos($job['packageName'], '/') === false) {
-                continue;
-            }
-
-            $listed[$job['packageName']] = true;
-            $packages[]                  = [$job['packageName'], $job['constraint']];
-        }
-
-        $this->container->get(ParallelDownloader::class)->download($packages, function ($packageName, $constraint) use (&$listed, &$packages, $pool): void {
-            /** @var \Composer\Package\PackageInterface $package */
-            foreach ($pool->whatProvides($packageName, $constraint, true) as $package) {
-                /** @var \Composer\Package\Link $link */
-                foreach (\array_merge($package->getRequires(), $package->getConflicts(), $package->getReplaces()) as $link) {
-                    if (isset($listed[$link->getTarget()]) || \strpos($link->getTarget(), '/') === false) {
-                        continue;
-                    }
-
-                    $listed[$link->getTarget()] = true;
-                    $packages[]                 = [$link->getTarget(), $link->getConstraint()];
-                }
-            }
-        });
-    }
-
-    /**
-     * Wrapper for the fetchAllFromOperations function.
-     *
-     * @see \Narrowspark\Automatic\Prefetcher\Prefetcher::fetchAllFromOperations()
-     *
-     * @param \Composer\Installer\InstallerEvent|\Composer\Installer\PackageEvent $event
-     *
-     * @return void
-     */
-    public function populateFilesCacheDir($event): void
-    {
-        $this->container->get(Prefetcher::class)->fetchAllFromOperations($event);
-    }
-
-    /**
-     * Adds the parallel downloader to composer.
-     *
-     * @param \Composer\Plugin\PreFileDownloadEvent $event
-     *
-     * @return void
-     */
-    public function onFileDownload(PreFileDownloadEvent $event): void
-    {
-        /** @var \Narrowspark\Automatic\Prefetcher\ParallelDownloader $rfs */
-        $rfs = $this->container->get(ParallelDownloader::class);
-
-        if ($event->getRemoteFilesystem() !== $rfs) {
-            $event->setRemoteFilesystem($rfs->setNextOptions($event->getRemoteFilesystem()->getOptions()));
         }
     }
