+++ src/Automatic/Container.php
@@ -93,7 +93,8 @@ final class Container implements ContainerContract
                     $container->get(IOInterface::class),
                     $container->get(Composer::class),
                     $container->get(Lock::class),
-                    $container->get(ClassFinder::class)
+                    $container->get(ClassFinder::class),
+                    $container
                 );
             },
             SkeletonInstaller::class => static function (ContainerContract $container) {
@@ -101,7 +102,8 @@ final class Container implements ContainerContract
                     $container->get(IOInterface::class),
                     $container->get(Composer::class),
                     $container->get(Lock::class),
-                    $container->get(ClassFinder::class)
+                    $container->get(ClassFinder::class),
+                    $container
                 );
             },
             ConfiguratorContract::class => static function (ContainerContract $container) {
