+++ src/Automatic/Prefetcher/Traits/PrefetcherTrait.php
@@ -1,0 +1,121 @@
+<?php
+declare(strict_types=1);
+namespace Narrowspark\Automatic\Prefetcher;
+
+use Composer\DependencyResolver\Pool;
+use Composer\Installer\InstallerEvent;
+use Composer\Plugin\PreFileDownloadEvent;
+use Composer\Repository\ComposerRepository as BaseComposerRepository;
+use Composer\Repository\RepositoryInterface;
+
+trait PrefetcherTrait
+{
+    /**
+     * Populate the provider cache.
+     *
+     * @param \Composer\Installer\InstallerEvent $event
+     *
+     * @return void
+     */
+    public function onPreDependenciesSolving(InstallerEvent $event): void
+    {
+        $listed   = [];
+        $packages = [];
+        $pool     = $event->getPool();
+        $pool     = \Closure::bind(function () {
+            foreach ($this->providerRepos as $k => $repo) {
+                $this->providerRepos[$k] = new class($repo) extends BaseComposerRepository {
+                    /**
+                     * A repository implementation.
+                     *
+                     * @var \Composer\Repository\RepositoryInterface
+                     */
+                    private $repo;
+
+                    /**
+                     * @param \Composer\Repository\RepositoryInterface $repo
+                     */
+                    public function __construct(RepositoryInterface $repo)
+                    {
+                        $this->repo = $repo;
+                    }
+
+                    /**
+                     * {@inheritdoc}
+                     */
+                    public function whatProvides(Pool $pool, $name, $bypassFilters = false)
+                    {
+                        $packages = [];
+
+                        if (! \method_exists($this->repo, 'whatProvides')) {
+                            return $packages;
+                        }
+
+                        foreach ($this->repo->whatProvides($pool, $name, $bypassFilters) as $k => $p) {
+                            $packages[$k] = clone $p;
+                        }
+
+                        return $packages;
+                    }
+                };
+            }
+
+            return $this;
+        }, clone $pool, $pool)();
+
+        foreach ($event->getRequest()->getJobs() as $job) {
+            if ($job['cmd'] !== 'install' || \strpos($job['packageName'], '/') === false) {
+                continue;
+            }
+
+            $listed[$job['packageName']] = true;
+            $packages[]                  = [$job['packageName'], $job['constraint']];
+        }
+
+        $this->container->get(ParallelDownloader::class)->download($packages, function ($packageName, $constraint) use (&$listed, &$packages, $pool): void {
+            /** @var \Composer\Package\PackageInterface $package */
+            foreach ($pool->whatProvides($packageName, $constraint, true) as $package) {
+                /** @var \Composer\Package\Link $link */
+                foreach (\array_merge($package->getRequires(), $package->getConflicts(), $package->getReplaces()) as $link) {
+                    if (isset($listed[$link->getTarget()]) || \strpos($link->getTarget(), '/') === false) {
+                        continue;
+                    }
+
+                    $listed[$link->getTarget()] = true;
+                    $packages[]                 = [$link->getTarget(), $link->getConstraint()];
+                }
+            }
+        });
+    }
+
+    /**
+     * Wrapper for the fetchAllFromOperations function.
+     *
+     * @see \Narrowspark\Automatic\Prefetcher\Prefetcher::fetchAllFromOperations()
+     *
+     * @param \Composer\Installer\InstallerEvent|\Composer\Installer\PackageEvent $event
+     *
+     * @return void
+     */
+    public function populateFilesCacheDir($event): void
+    {
+        $this->container->get(Prefetcher::class)->fetchAllFromOperations($event);
+    }
+
+    /**
+     * Adds the parallel downloader to composer.
+     *
+     * @param \Composer\Plugin\PreFileDownloadEvent $event
+     *
+     * @return void
+     */
+    public function onFileDownload(PreFileDownloadEvent $event): void
+    {
+        /** @var \Narrowspark\Automatic\Prefetcher\ParallelDownloader $rfs */
+        $rfs = $this->container->get(ParallelDownloader::class);
+
+        if ($event->getRemoteFilesystem() !== $rfs) {
+            $event->setRemoteFilesystem($rfs->setNextOptions($event->getRemoteFilesystem()->getOptions()));
+        }
+    }
+}
